% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemiMyocardialInfarction.R
\docType{package}
\name{stemiMyocardialInfarction-package}
\alias{stemiMyocardialInfarction}
\alias{stemiMyocardialInfarction-package}
\title{stemiMyocardialInfarction: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
